+++ demos/address_verify.php
@@ -1,12 +1,16 @@
 <?php
+// USAGE: address_verify.php USPS_USER [USPS_PASSWORD]
+// NOTE:  Be sure to run composer dump-autoload to update autoload file mapping
+require __DIR__ . '/../vendor/autoload.php';
+use USPS\{USPSAddress,USPSAddressVerify};
 
-// Load the class
-require_once('../USPSAddressVerify.php');
 // Initiate and set the username provided from usps
-$verify = new USPSAddressVerify('xxxx');
+$user   = $argv[1] ?? $_GET['user'] ?? 'xxx';
+$pass   = $argv[2] ?? $_GET['password'] ?? '';
+$verify = new USPSAddressVerify($user, $pass);
 
 // During test mode this seems not to always work as expected
-//$verify->setTestMode(true);
+$verify->setTestMode(true);
 
 // Create new address object and assign the properties
 // apartently the order you assign them is important so make sure
