+++ demos/domestic_rate.php
@@ -1,12 +1,14 @@
 <?php
-// Load the class
-require_once('../USPSRate.php');
+// Load the class; use Composer autoloader
+// NOTE:  Be sure to run composer dump-autoload to update autoload file mapping
+require __DIR__ . '/../vendor/autoload.php';
+use USPS\{USPSRate,USPSRatePackage};
 
 // Initiate and set the username provided from usps
 $rate = new USPSRate('xxxx');
 
 // During test mode this seems not to always work as expected
-//$rate->setTestMode(true);
+$rate->setTestMode(true);
 
 // Create new package object and assign the properties
 // apartently the order you assign them is important so make sure
