+++ demos/firstclassservicestandard.php
@@ -1,9 +1,12 @@
 <?php
-// Load the class
-require_once('../USPSFirstClassServiceStandards.php');
+// USAGE: firstclassservicestandard.php USPS_USER
+// NOTE:  Be sure to run composer dump-autoload to update autoload file mapping
+require __DIR__ . '/../vendor/autoload.php';
+use USPS\USPSFirstClassServiceStandards;
 
 // Initiate and set the username provided from usps.
-$delivery = new USPSFirstClassServiceStandards('xxxx');
+$user   = $argv[1] ?? $_GET['user'] ?? 'xxx';
+$delivery = new USPSFirstClassServiceStandards($user);
 
 // During test mode this seems not to always work as expected.
 $delivery->setTestMode(true);
