+++ demos/priority_label.php
@@ -1,9 +1,14 @@
 <?php
+// USAGE: priority_label.php USPS_USER USPS_PASSWORD
+// NOTE: APIs that return printed labels or barcodes become available after applying for advanced permissions from the USPS APIs Functional Team.
+//       Be sure to run composer dump-autoload to update autoload file mapping
+require __DIR__ . '/../vendor/autoload.php';
+use USPS\USPSPriorityLabel;
 
-// Load the class
-require_once('../USPSPriorityLabel.php');
 // Initiate and set the username provided from usps
-$label = new USPSPriorityLabel('xxxx');
+$user   = $argv[1] ?? $_GET['user'] ?? 'xxx';
+$pass   = $argv[2] ?? $_GET['password'] ?? '';
+$label = new USPSPriorityLabel($user, $pass);
 
 // During test mode this seems not to always work as expected
 $label->setTestMode(true);
@@ -29,7 +34,7 @@ if($label->isSuccess()) {
   //echo "\n Confirmation:" . $label->getConfirmationNumber();
 
   $label = $label->getLabelContents();
-  
+
   if($label) {
   	$contents = base64_decode($label);
   	header('Content-type: application/pdf');
